Add HasConfig and HasView traits
