+++ src/Views/Filter.php
@@ -8,6 +8,8 @@ use Rappasoft\LaravelLivewireTables\Views\Traits\IsFilter;
 abstract class Filter
 {
     use IsFilter;
+    
+    protected string $view = '';
 
     public function __construct(string $name, ?string $key = null)
     {
