+++ src/Views/Traits/HasConfig.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Views\Traits;
+
+use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
+use Rappasoft\LaravelLivewireTables\Views\{Column,Filter};
+
+trait HasConfig
+{
+    public array $config = [];
+
+    /**
+     * @param  array<mixed>  $config
+     */
+    public function config(array $config = []): Filter
+    {
+        $this->config = $config;
+
+        return $this;
+    }
+
+
+    /**
+     * Get the filter configs.
+     *
+     * @return array<mixed>
+     */
+    public function getConfigs(): array
+    {
+        return $this->config;
+    }
+
+    /**
+     * Get a single filter config.
+     *
+     * @return mixed
+     */
+    public function getConfig(string $key)
+    {
+        return $this->config[$key] ?? null;
+    }
+
+    public function hasConfigs(): bool
+    {
+        return count($this->getConfigs()) > 0;
+    }
+
+    public function hasConfig(string $key): bool
+    {
+        return array_key_exists($key, $this->getConfigs()) && $this->getConfig($key) !== null;
+    }
+}
