+++ src/Views/Traits/Helpers/FilterHelpers.php
@@ -23,26 +23,6 @@ trait FilterHelpers
     }
 
     /**
-     * Get the filter configs.
-     *
-     * @return array<mixed>
-     */
-    public function getConfigs(): array
-    {
-        return $this->config;
-    }
-
-    /**
-     * Get a single filter config.
-     *
-     * @return mixed
-     */
-    public function getConfig(string $key)
-    {
-        return $this->config[$key] ?? null;
-    }
-
-    /**
      * Get the filter keys.
      *
      * @return array<mixed>
@@ -108,15 +88,6 @@ trait FilterHelpers
         return $this->getCustomFilterPillValues()[$value] ?? null;
     }
 
-    public function hasConfigs(): bool
-    {
-        return count($this->getConfigs()) > 0;
-    }
-
-    public function hasConfig(string $key): bool
-    {
-        return array_key_exists($key, $this->getConfigs()) && $this->getConfig($key) !== null;
-    }
 
     public function isHiddenFromMenus(): bool
     {
